Add files via upload
Project Overview

In this project we are building a  recurrent neural network for the purpose of determining the sentiment of a movie review using the IMDB data set. We  create this model using Amazon's SageMaker service. In addition, we  deployed your model and construct a simple web app which will interact with the deployed model.